PraxisSoftware/requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg[binary]>=3.1,<4.0
dj-database-url>=2.1,<3.0

# CORS
django-cors-headers>=4.3,<5.0

# Environment
python-dotenv>=1.0,<2.0

# Production Server
gunicorn>=21.0,<23.0
whitenoise>=6.6,<7.0

# Background Tasks (optional)
celery>=5.3,<6.0
redis>=5.0,<6.0

# Development
pytest>=7.4,<9.0
pytest-django>=4.7,<5.0

# Formatting / Linting
black>=24.0,<26.0
isort>=5.13,<6.0
ruff>=0.6,<1.0
